chore: move getTransactionCount tests to batch3
