+++ packages/server/tests/acceptance/index.spec.ts
@@ -137,9 +137,9 @@ describe('RPC Server Acceptance Tests', function () {
 
     function runLocalHederaNetwork() {
         // set env variables for docker images until local-node is updated
-        process.env['NETWORK_NODE_IMAGE_TAG'] = '0.40.0-alpha.1';
-        process.env['HAVEGED_IMAGE_TAG'] = '0.40.0-alpha.1';
-        process.env['MIRROR_IMAGE_TAG'] = '0.84.0-rc2';
+        process.env['NETWORK_NODE_IMAGE_TAG'] = '0.40.1';
+        process.env['HAVEGED_IMAGE_TAG'] = '0.40.1';
+        process.env['MIRROR_IMAGE_TAG'] = '0.84.0';
 
         console.log(`Docker container versions, services: ${process.env['NETWORK_NODE_IMAGE_TAG']}, mirror: ${process.env['MIRROR_IMAGE_TAG']}`);
