+++ packages/server/tests/contracts/Deployer.json
@@ -1,0 +1,117 @@
+{
+  "_format": "hh-sol-artifact-1",
+  "contractName": "Deployer",
+  "sourceName": "contracts/Deployer.sol",
+  "abi": [
+    {
+      "inputs": [],
+      "stateMutability": "payable",
+      "type": "constructor"
+    },
+    {
+      "inputs": [],
+      "name": "counter",
+      "outputs": [
+        {
+          "internalType": "uint256",
+          "name": "",
+          "type": "uint256"
+        }
+      ],
+      "stateMutability": "view",
+      "type": "function"
+    },
+    {
+      "inputs": [],
+      "name": "deployViaCreate",
+      "outputs": [
+        {
+          "internalType": "address",
+          "name": "",
+          "type": "address"
+        }
+      ],
+      "stateMutability": "nonpayable",
+      "type": "function"
+    },
+    {
+      "inputs": [],
+      "name": "deployViaCreate2",
+      "outputs": [
+        {
+          "internalType": "address",
+          "name": "",
+          "type": "address"
+        }
+      ],
+      "stateMutability": "nonpayable",
+      "type": "function"
+    },
+    {
+      "inputs": [],
+      "name": "getMockContractAddress",
+      "outputs": [
+        {
+          "internalType": "address",
+          "name": "",
+          "type": "address"
+        }
+      ],
+      "stateMutability": "view",
+      "type": "function"
+    },
+    {
+      "inputs": [],
+      "name": "salt",
+      "outputs": [
+        {
+          "internalType": "uint256",
+          "name": "",
+          "type": "uint256"
+        }
+      ],
+      "stateMutability": "view",
+      "type": "function"
+    },
+    {
+      "inputs": [],
+      "name": "salt2",
+      "outputs": [
+        {
+          "internalType": "uint256",
+          "name": "",
+          "type": "uint256"
+        }
+      ],
+      "stateMutability": "view",
+      "type": "function"
+    },
+    {
+      "inputs": [
+        {
+          "internalType": "uint256",
+          "name": "_counter",
+          "type": "uint256"
+        }
+      ],
+      "name": "updateCounter",
+      "outputs": [
+        {
+          "internalType": "uint256",
+          "name": "",
+          "type": "uint256"
+        }
+      ],
+      "stateMutability": "nonpayable",
+      "type": "function"
+    },
+    {
+      "stateMutability": "payable",
+      "type": "receive"
+    }
+  ],
+  "bytecode": "0x60806040526001600055600180556002805560405161001d9061005f565b604051809103906000f080158015610039573d6000803e3d6000fd5b50600380546001600160a01b0319166001600160a01b039290921691909117905561006b565b60a58061034f83390190565b6102d58061007a6000396000f3fe6080604052600436106100745760003560e01c80636e6662b91161004e5780636e6662b91461010b578063bfa0b13314610120578063c648049d14610136578063dbb6f04a1461015957600080fd5b806314a7862c1461008057806331bf722a146100d157806361bc221a146100f557600080fd5b3661007b57005b600080fd5b34801561008c57600080fd5b5060035473ffffffffffffffffffffffffffffffffffffffff165b60405173ffffffffffffffffffffffffffffffffffffffff90911681526020015b60405180910390f35b3480156100dd57600080fd5b506100e760025481565b6040519081526020016100c8565b34801561010157600080fd5b506100e760005481565b34801561011757600080fd5b506100a761016e565b34801561012c57600080fd5b506100e760015481565b34801561014257600080fd5b506100e76101513660046101e1565b600081905590565b34801561016557600080fd5b506100a76101a0565b60008060405161017d906101d5565b604051809103906000f080158015610199573d6000803e3d6000fd5b5092915050565b60008060005460001b6040516101b5906101d5565b8190604051809103906000f5905080158015610199573d6000803e3d6000fd5b60a5806101fb83390190565b6000602082840312156101f357600080fd5b503591905056fe6080604052348015600f57600080fd5b5060878061001e6000396000f3fe6080604052348015600f57600080fd5b506004361060325760003560e01c806338cc483114603757806383197ef014604a575b600080fd5b6040805130815290519081900360200190f35b604f33ff5b00fea2646970667358221220e4694c5da062c0f5d60f55b135a5e3c1f3741556180ff1dd7fb1f1658265e42164736f6c63430008090033a2646970667358221220dc5790d48165edb918974a228dda59d5f9a307589b9d442a8ed826a61484f4b464736f6c634300080900336080604052348015600f57600080fd5b5060878061001e6000396000f3fe6080604052348015600f57600080fd5b506004361060325760003560e01c806338cc483114603757806383197ef014604a575b600080fd5b6040805130815290519081900360200190f35b604f33ff5b00fea2646970667358221220e4694c5da062c0f5d60f55b135a5e3c1f3741556180ff1dd7fb1f1658265e42164736f6c63430008090033",
+  "deployedBytecode": "0x6080604052600436106100745760003560e01c80636e6662b91161004e5780636e6662b91461010b578063bfa0b13314610120578063c648049d14610136578063dbb6f04a1461015957600080fd5b806314a7862c1461008057806331bf722a146100d157806361bc221a146100f557600080fd5b3661007b57005b600080fd5b34801561008c57600080fd5b5060035473ffffffffffffffffffffffffffffffffffffffff165b60405173ffffffffffffffffffffffffffffffffffffffff90911681526020015b60405180910390f35b3480156100dd57600080fd5b506100e760025481565b6040519081526020016100c8565b34801561010157600080fd5b506100e760005481565b34801561011757600080fd5b506100a761016e565b34801561012c57600080fd5b506100e760015481565b34801561014257600080fd5b506100e76101513660046101e1565b600081905590565b34801561016557600080fd5b506100a76101a0565b60008060405161017d906101d5565b604051809103906000f080158015610199573d6000803e3d6000fd5b5092915050565b60008060005460001b6040516101b5906101d5565b8190604051809103906000f5905080158015610199573d6000803e3d6000fd5b60a5806101fb83390190565b6000602082840312156101f357600080fd5b503591905056fe6080604052348015600f57600080fd5b5060878061001e6000396000f3fe6080604052348015600f57600080fd5b506004361060325760003560e01c806338cc483114603757806383197ef014604a575b600080fd5b6040805130815290519081900360200190f35b604f33ff5b00fea2646970667358221220e4694c5da062c0f5d60f55b135a5e3c1f3741556180ff1dd7fb1f1658265e42164736f6c63430008090033a2646970667358221220dc5790d48165edb918974a228dda59d5f9a307589b9d442a8ed826a61484f4b464736f6c63430008090033",
+  "linkReferences": {},
+  "deployedLinkReferences": {}
+}
