+++ packages/server/tests/helpers/prerequisite.ts
@@ -11,9 +11,9 @@ const RELAY_URL = process.env.E2E_RELAY_HOST || LOCAL_RELAY_URL;
 
 (function () {
   if (USE_LOCAL_NODE) {
-    process.env['NETWORK_NODE_IMAGE_TAG'] = '0.40.0-alpha.1';
-    process.env['HAVEGED_IMAGE_TAG'] = '0.40.0-alpha.1';
-    process.env['MIRROR_IMAGE_TAG'] = '0.82.1';
+    process.env['NETWORK_NODE_IMAGE_TAG'] = '0.40.1';
+    process.env['HAVEGED_IMAGE_TAG'] = '0.40.1';
+    process.env['MIRROR_IMAGE_TAG'] = '0.84.0';
 
     console.log(`Docker container versions, services: ${process.env['NETWORK_NODE_IMAGE_TAG']}, mirror: ${process.env['MIRROR_IMAGE_TAG']}`);
